index.rst: reword the heading of the main TOC
I want to add a `Glossary` to the TOC, and so changing "Getting
Involved" to a more general "Getting Started with ZeroVM" for other
types of general information.

For example, in the future I also want to add tutorials, installation
guides, etc., and such topic make more sense under the new heading.