contributing: shuffle headings around
The "Contributing Guidelines" section is now more a section about how
to write good commits, so it has been renamed to better match its
content.

This means that we now have main headings about how to use pull
requests, how to write good commits and how to use the long-term
branches in the repositories.