contributing: remove C4 reference
Instead of referring to a large set of rules in an external document,
we now have sufficient guidelines in our own document to help
contributors. Our document emphasizes good code quality.

The rules and guidelines in our document are generic and have been
inspired by other large open source projects such as Mercurial, Git,
and OpenStack:

  http://mercurial.selenic.com/wiki/ContributingChanges
  https://github.com/git/git/blob/master/Documentation/SubmittingPatches
  https://wiki.openstack.org/wiki/GitCommitMessages

The idea is that this process should be familiar to other
contributors, while ensuring that we maintain high code and design
quality.