snakebin: mv OS_STORAGE_TOKEN OS_STORAGE_ACCOUNT
OS_STORAGE_TOKEN doesn't seem like the right name, and may be confused
with the X-Storage-Token header (similar in use to the X-Auth-Token
header submitted with Swift HTTP requests). (See
http://docs.openstack.org/developer/swift/overview_auth.html for more
information about X-Storage-Token.)

Also, the data this variable represents is not a storage token; it's the
ID of the storage account. Therefore, OS_STORAGE_ACCOUNT is a better
name.